+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Adheres to [Semantic Versioning](http://semver.org/).
 * Support EPSG 900913 (GOOGLE)
 * tiff-ios version updated to 1.0.1
 * proj4-ios verion 4.9.3
+* Bug fixes for elevation extension when handling null elevations
 
 ## [1.2.1](https://github.com/ngageoint/geopackage-ios/releases/tag/1.2.1) (02-02-2017)